Added time measurement features
Drug the functions you want to disable into the DISABLE_THREAD and keep the one for testing. Skip the vTaskDelayUntil() and modify the function so that a key press message is generated for each of the 12 keys on every iteration of the function, regardless of which keys are actually pressed.